requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
openai
numpy
python-multipart
pdf2image
pytesseract